Changes to be committed: 	modified:   lib/controller/active_members_controller.dart 	modified:   lib/controller/members_controller.dart 	modified:   lib/models/members_model.dart

+++ lib/controller/active_members_controller.dart
@@ -5,7 +5,7 @@ class ActiveMembersController {
 
   void addActiveMember({
     required String nama,
-    required double noTelepon,
+    required int noTelepon,
     required String jabatan,
     required String alamat,
   }) {
